fix layout & components
