+++ src/resources/js/Pages/Profile/Edit.jsx
@@ -1,40 +1,29 @@
-import AuthenticatedLayout from "@/Layouts/AuthenticatedLayout";
 import DeleteUserForm from "./Partials/DeleteUserForm";
 import UpdatePasswordForm from "./Partials/UpdatePasswordForm";
-import UpdateProfileInformationForm from "./ProfileEdit";
-import { Head } from "@inertiajs/react";
+import ProfileEdit from "./Partials/ProfileEdit";
+import GeneralLayout from "@/Layouts/GeneralLayout";
 
-export default function Edit({ auth, mustVerifyEmail, status }) {
+export default function Edit({ auth, mustVerifyEmail, status, errors, flash }) {
     return (
-        <AuthenticatedLayout
-            user={auth.user}
-            header={
-                <h2 className="font-semibold text-xl text-gray-800 leading-tight">
-                    Profile
-                </h2>
-            }
-        >
-            <Head title="Profile" />
-
-            <div className="py-12">
-                <div className="max-w-7xl mx-auto sm:px-6 lg:px-8 space-y-6">
-                    <div className="p-4 sm:p-8 bg-white shadow sm:rounded-lg">
-                        <UpdateProfileInformationForm
-                            mustVerifyEmail={mustVerifyEmail}
-                            status={status}
-                            className="max-w-xl"
-                        />
-                    </div>
-
-                    <div className="p-4 sm:p-8 bg-white shadow sm:rounded-lg">
-                        <UpdatePasswordForm className="max-w-xl" />
-                    </div>
-
-                    <div className="p-4 sm:p-8 bg-white shadow sm:rounded-lg">
-                        <DeleteUserForm className="max-w-xl" />
-                    </div>
-                </div>
+        <GeneralLayout auth={auth}>
+            <div className="pb-10">
+                <ProfileEdit
+                    auth={auth}
+                    mustVerifyEmail={mustVerifyEmail}
+                    status={status}
+                    errors={errors}
+                    flash={flash}
+                    className="max-w-xl"
+                />
             </div>
-        </AuthenticatedLayout>
+
+            <div className="p-4 sm:p-8 bg-gray-100 shadow sm:rounded-lg">
+                <UpdatePasswordForm className="max-w-xl" />
+            </div>
+
+            <div className="p-4 sm:p-8 bg-white shadow sm:rounded-lg">
+                <DeleteUserForm className="max-w-xl" />
+            </div>
+        </GeneralLayout>
     );
 }
