+++ src/routes/web.php
@@ -34,11 +34,6 @@ Route::middleware('auth')->group(function () {
     Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
     Route::post('/profile', [ProfileController::class, 'update'])->name('profile.update');
     Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
-
-
-    Route::get('/user/delete', function () {
-        return Inertia::render('Profile/DeleteUserForm');
-    });
 });
 
 // Participants
